changed: Removed self-built boost. Use the libs from MinGW Distro instead. removed: Removed self-made Pointer classes. Use std pointers supported by C++11 instead. changed: Make String no longer a descendant of std::wstring. added: Added many utility function to String.

+++ trunk/components/common/Object.h
@@ -15,7 +15,7 @@ namespace viu2x {
             Object();
             virtual ~Object();
 
-            virtual uint32_t getHash();
+            virtual uint64_t getHash();
     };
 
 }
